+++ src/finance/include/finance/cash.hpp
@@ -45,6 +45,9 @@ namespace finance
         [[nodiscard]] micro_units getAmount() const;
         [[nodiscard]] Currency    getCurrency() const;
         [[nodiscard]] std::string toString() const;
+        [[nodiscard]] std::string toString(
+            std::optional<std::uint8_t> nDecimalPlaces
+        ) const;
     };
 
 }   // namespace finance